Use Karma for running the polyfill tests
