+++ packages/polyfill-library/README.md
@@ -100,6 +100,12 @@ Development of polyfill-library happens on GitHub. Read below to learn how you c
 
 Read our [contributing guide][contributing-guide] to learn about our development process, how to propose bugfixes and improvements, and how to build and test your changes.
 
+```
+npm run test-all-polyfills # Run the tests for all polyfills using BrowserStack
+npm run test-polyfills -- --feature=Array.from # Run the tests for Array.from
+npm run test-polyfills -- --feature=Array.from --browserstack # Run the tests for Array.from using BrowserStack
+```
+
 ### License
 
 Polyfill-library is [MIT licensed][license].
