+++ packages/polyfill-library/package.json
@@ -27,7 +27,8 @@
     "clean": "rimraf ./polyfills/__dist",
     "build": "npm run clean && node tasks/node/updatesources && node tasks/node/buildsources",
     "fmt": "eslint . --fix",
-    "test-polyfills": "node test/polyfills/test-multiple-features.js",
+    "test-polyfills": "karma start config.js",
+    "test-all-polyfills": "node ./test/polyfills/test-all-features.js",
     "test-node": "mocha test/node/**/*.js",
     "test-node-unit": "nyc mocha test/unit --recursive",
     "test-node-coverage": "nyc check-coverage --lines 90 --branches 90 --functions 90",
@@ -80,6 +81,7 @@
   "devDependencies": {
     "eslint": "^4.9.0",
     "execa": "^1.0.0",
+    "globby": "^8.0.1",
     "karma": "^3.1.1",
     "karma-browserstack-launcher": "^1.3.0",
     "karma-mocha": "^1.3.0",
